requirements.txt
arabic-reshaper==2.1.0
asgiref==3.2.10
backports.zoneinfo==0.2.1
Django==3.1
django-filter==21.1
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
Pillow==8.1.0
PyPDF2==1.26.0
python-bidi==0.4.2
pytz==2020.1
reportlab==3.5.59
six==1.15.0
sqlparse==0.3.1
waitress==2.0.0
webencodings==0.5.1
whitenoise==6.0.0
xhtml2pdf==0.2.5
